+++ package.json
@@ -38,10 +38,10 @@
     "broccoli-string-replace": "~0.0.2",
     "broccoli-uglify-js": "~0.1.3",
     "chalk": "~0.4.0",
-    "ember-cli": "^0.2.7",
+    "ember-cli": "^1.13.8",
     "ember-cli-release": "^0.2.2",
     "ember-cli-sauce": "^1.3.0",
-    "emberjs-build": "0.2.1",
+    "emberjs-build": "0.4.5",
     "git-repo-version": "^0.1.2",
     "handlebars": "^3.0.2",
     "morph-range": "0.2.5",
